confirm first argument is directory and exist

+++ bash/clean.sh
@@ -13,11 +13,17 @@ then
     usage
     exit
 fi
+if [ ! -d "$1" ]
+then
+    echo "Directory $1 does not exist!"
+    exit
+fi
+SOURCE_DIR="$1"
 
 TMP="/tmp"
-find $1/resource/file -type f \( -name "*.pdf" -or -name "*.jpg" -or -name "*.swf" -or -name "*.mp3" -or -name "*.png" \) | sed "s:$1/resource/::" > $TMP/all
-grep -o --no-filename 'file[^"]*\.\(pdf\|jpg\|swf\|mp3\)' $1/json/*.json| uniq | sed -e 's:\\/:/:g' >  $TMP/used
-fgrep -v $TMP/all -f $TMP/used
+find $SOURCE_DIR/resource/file -type f \( -name "*.pdf" -or -name "*.jpg" -or -name "*.swf" -or -name "*.mp3" -or -name "*.png" \) | sed "s:$SOURCE_DIR/resource/::" > $TMP/all
+grep -o --no-filename 'file[^"]*\.\(pdf\|jpg\|swf\|mp3\)' $SOURCE_DIR/json/*.json| uniq | sed -e 's:\\/:/:g' >  $TMP/used
+fgrep -v $TMP/all -f $TMP/used > $TMP/unused
 rm $TMP/used
 rm $TMP/all
